Refactor message storage, add caching
