--- src/message-list/messagesSelectors.js
@@ -1,2 +1,0 @@
-export const getMessages = (state) =>
-  state.messages;
