+++ src/reducers.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import account from './account/accountReducers';
 import app from './app/appReducers';
 import events from './events/eventReducers';
-import messages from './message-list/messagesReducers';
 import nav from './nav/navReducers';
 import realm from './realm/realmReducers';
 import subscriptions from './subscriptions/subscriptionsReducers';
@@ -14,7 +13,6 @@ export default combineReducers({
   app,
   chat,
   events,
-  messages,
   nav,
   realm,
   subscriptions,
